+++ src/extension/actions.js
@@ -29,6 +29,7 @@ import { ADMIN_ORIGIN, createAdminUrl } from './utils/admin.js';
 import { getConfig } from './config.js';
 import { getDisplay, setDisplay } from './display.js';
 import { urlCache } from './url-cache.js';
+import { saveDocument } from './sharepoint.js';
 
 /**
  * Updates the auth token via external messaging API (admin only).
@@ -301,8 +302,7 @@ async function login({
  */
 async function addRemoveProject(tab) {
   const matches = await getProjectMatches(await getProjects(), tab);
-  const config = matches.length === 1 && !matches[0].transient
-    ? matches[0] : await getProjectFromUrl(tab);
+  const config = matches.length === 1 ? matches[0] : await getProjectFromUrl(tab);
 
   await showSidekickIfHidden();
   if (isValidProject(config)) {
@@ -538,6 +538,7 @@ export const internalActions = {
   getProfilePicture,
   guessAEMSite,
   updateProject,
+  saveDocument,
 };
 
 /**