+++ src/extension/app/store/bulk.js
@@ -163,16 +163,26 @@ export class BulkStore {
         && !row.querySelector('svg')?.parentElement.className.toLowerCase().includes('folder'))
       // extract file name and type
       .map((row) => {
-        const info = row.getAttribute('aria-label') || row.querySelector('span')?.textContent;
+        let file = row
+          .querySelector('[data-id="heroField"], [data-automationid="name"]') // list, grid
+          ?.textContent.trim();
+        let type = (file || '').split('.').pop();
 
-        // info format: bla.docx, docx File, Private, Modified 8/28/2023, edited by Jane, 1 KB
-        let type = info.match(/, ([\p{L}\p{N}]+) [\p{L}\p{N}]+,/u)?.[1];
-        let file = type && info.split(`, ${type}`)[0];
+        // istanbul ignore next 13
+        if (!file) {
+          // fallback to previous SP DOM version and retrieve type and file from info string
+          const info = row.getAttribute('aria-label') || row.querySelector('span')?.textContent || '';
 
-        if (!type) {
-          type = info.split('.').pop();
-          file = info;
+          // info format: bla.docx, docx File, Private, Modified 8/28/2023, edited by Jane, 1 KB
+          type = info.match(/, ([\p{L}\p{N}]+) [\p{L}\p{N}]+,/u)?.[1];
+          file = type && info.split(`, ${type}`)[0];
+
+          if (!type) {
+            type = info.split('.').pop();
+            file = info;
+          }
         }
+
         return {
           file,
           type,
@@ -190,15 +200,21 @@ export class BulkStore {
    * @returns {BulkSelection} The selection
    */
   #getGoogleDriveBulkSelection(document) {
-    return [...document.querySelectorAll('#drive_main_page [role="row"][aria-selected="true"]')]
+    return [...document.querySelectorAll('#drive_main_page [aria-selected="true"]')]
       // extract file name and type
       .map((row) => {
-        const file = (row.querySelector(':scope div[role="gridcell"] > div > div:nth-child(2) div[jsname]') // grid layout
-          || row.querySelector(':scope div[role="gridcell"] > div:nth-of-type(2)')) // list layout
+        const file = (row.querySelector(':scope td div[data-id] > span > strong') // list layout
+          || row.querySelector(':scope div > div:nth-child(2) > div:nth-child(2) > div') // grid layout
+          // istanbul ignore next 2
+          || row.querySelector(':scope div[role="gridcell"] > div > div:nth-child(2) div[jsname]') // legacy grid layout
+          || row.querySelector(':scope div[role="gridcell"] > div:nth-of-type(2) > div:not([role="button"])')) // legacy list layout
           ?.textContent.trim();
 
         // use path in icon svg to determine type
-        const typeHint = row.querySelector(':scope div[role="gridcell"] > div path') // list & grid layout
+        const typeHint = (row.querySelector(':scope td div > svg > path') // list layout
+          || row.querySelector(':scope div > div:nth-child(2) > div:nth-child(1) div > svg > path')
+          // istanbul ignore next
+          || row.querySelector(':scope div[role="gridcell"] > div path')) // legacy list & grid layout
           ?.getAttribute('d').slice(-4);
         let type = 'unknown';
         if (typeHint) {