+++ test/app/store/bulk.test.js
@@ -363,7 +363,7 @@ describe('Test Bulk Store', () => {
         await waitUntil(() => startJobStub.calledWith('preview', [
           '/foo/document',
           '/foo/spreadsheet.json',
-          '/foo/other',
+          '/foo/other.unknown',
         ]));
         await waitUntil(() => getJobStub.calledWith('preview', '123'), null, { timeout: 2000 });
         expect(bulkStore.progress.processed).to.equal(1);
@@ -1150,7 +1150,7 @@ describe('Test Bulk Store', () => {
 
       // delete file icon to simulate unknown file type
       sidekickTest.toggleAdminItems(['document']);
-      sidekickTest.bulkRoot.querySelector('div.file[aria-selected="true"] svg').remove();
+      sidekickTest.bulkRoot.querySelector('.file[aria-selected="true"] svg').remove();
       bulkStore.initStore(appStore.location);
       await waitUntil(() => bulkStore.selection.length === 1);
 
@@ -1169,7 +1169,7 @@ describe('Test Bulk Store', () => {
 
       // add .docx extension to file name
       sidekickTest.bulkRoot
-        .querySelector('div#file-gdoc div[data-tooltip]')
+        .querySelector('#file-gdoc strong')
         .textContent = 'document.docx';
       sidekickTest.toggleAdminItems(['document']);
 
@@ -1195,7 +1195,7 @@ describe('Test Bulk Store', () => {
 
       // add .xlsx extension to file name
       sidekickTest.bulkRoot
-        .querySelector('div#file-gsheet div[data-tooltip]')
+        .querySelector('#file-gsheet strong')
         .textContent = 'spreadsheet.xlsx';
       sidekickTest.toggleAdminItems(['spreadsheet']);
 
@@ -1207,6 +1207,25 @@ describe('Test Bulk Store', () => {
 
       await waitUntil(() => updateStub.called);
       expect(updateStub.calledWith('/spreadsheet.xlsx')).to.be.true;
+    });
+
+    it('missing file name in sharepoint', async () => {
+      // mock sharepoint
+      sidekickTest
+        .mockLocation(getAdminLocation(HelixMockContentSources.SHAREPOINT))
+        .mockAdminDOM(HelixMockContentSources.SHAREPOINT);
+      await appStore.loadContext(sidekickTest.createSidekick(), sidekickTest.config);
+      await aTimeout(500);
+
+      // remove file name from first file
+      sidekickTest.bulkRoot.querySelector('[data-id="heroField"]').remove();
+      // select first file
+      sidekickTest.bulkRoot.querySelector('.file').setAttribute('aria-selected', 'true');
+
+      bulkStore.initStore(appStore.location);
+      await aTimeout(500);
+
+      expect(bulkStore.selection.length).to.equal(0);
     });
 
     it('rejects illegal folder path', async () => {